+++ app/src/Components/Event/EventSearchQuickSearch.tsx
@@ -52,7 +52,7 @@ interface EventSearchQuickSearchProps {
  * @param {EventQuickSearchChoice[]} props.choices The choices which will be used for the buttons on the quick search
  * @param {EventsSearchField} props.field The field to search on
  * @param {string} props.heading Text for the heading of this quick search group
- * @returns ReactElement Component
+ * @returns {React.ReactElement} Component
  */
 const EventSearchQuickSearchButtons: FC<EventSearchQuickSearchProps> = ({
   choices,
@@ -67,7 +67,7 @@ const EventSearchQuickSearchButtons: FC<EventSearchQuickSearchProps> = ({
     (state: { events: EventsState }) => state.events.upcoming,
   )
 
-  const allEvents = [...allUpcomingEvents, ...allPastEvents]
+  const allEvents = [...(allUpcomingEvents || []), ...(allPastEvents || [])]
 
   const handleSearch = (quickSearchChoice: string): void => {
     const eventsSearchResults = searchByArray(allEvents, field, [